replacing old home page button with new one
