+++ frontend/src/components/pages/ComponentLibrary.tsx
@@ -8,7 +8,6 @@ import {
 } from "@chakra-ui/react";
 
 import { ArrowBackOutlineIcon, ArrowForwardOutlineIcon } from "../common/icons";
-import MainPageButton from "../common/MainPageButton";
 import AdminConfirmationMessage from "../common/Admin/AdminConfirmationMessage";
 import RemoveUserPopover from "../common/Admin/RemoveUserPopover";
 import QuestionCard from "../common/QuestionCard";
@@ -194,7 +193,6 @@ const ComponentLibrary = (): React.ReactElement => {
       <ButtonExamples />
       <AdminConfirmationMessage />
       <RemoveUserPopover name="Sam Hutchinson" email="shutch@jumpmath.ca" />
-      <MainPageButton />
       <QuestionCard
         date={new Date()}
         title="Question Title"
