+++ frontend/src/components/pages/StudentPage.tsx
@@ -1,11 +1,9 @@
 import React from "react";
-import MainPageButton from "../common/MainPageButton";
 
 const StudentPage = (): React.ReactElement => {
   return (
     <div style={{ textAlign: "center", width: "25%", margin: "0px auto" }}>
       <h1>Student Page</h1>
-      <MainPageButton />
     </div>
   );
 };
